requirements.txt
ansible==2.8.2
asn1crypto==0.24.0
bcrypt==3.1.6
boto==2.49.0
boto3==1.9.130
botocore==1.12.130
cffi==1.12.2
cryptography==2.6.1
cycler==0.10.0
decorator==4.4.0
docutils==0.14
html5lib==1.0.1
Jinja2==2.10.1
jmespath==0.9.4
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.0
mechanize==0.4.2
networkx==2.3
numpy==1.16.4
paramiko==2.4.2
pyasn1==0.4.5
pycparser==2.19
Pygments==2.3.1
PyNaCl==1.3.0
pyparsing==2.4.0
python-dateutil==2.8.0
PyYAML==5.1
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.2
webencodings==0.5.1
